feat(*): optionally set last tag explicitly
For most workflows the task is smart enough
to figure out the last tag on it's own.

However for workflows which create tags
as leafs the auto recognition does not work.

E.g.

      0.3.0
   |/
   |  0.2.0
   |/
   |  0.1.0
   |/

Such a workflow makes sense if tags should
include a dist folder wheras the development
branch should not.

With this change one can set the last tag
explicitly with the `lastTag` option.