feat(*): Allow setting the cmd which is used to figure out previous tag
For most workflows the task is smart enough
to figure out the last tag on it's own.

However for workflows which create tags
as leafs the auto recognition does not work.

E.g.

      0.3.0               0.3.0
   |/                       |
   |  0.2.0  rather than  0.2.0
   |/                       |
   |  0.1.0               0.1.0
   |/                       |

Such a workflow makes sense if tags should
include a dist folder wheras the development
branch should not.

With this change one can set the cmd which is
used to figure out the previous tag.

For instance to match the sketched scenario
one could set it to:

git tag | sort -n -t. -k1,1 -k2,2 -k3,3 | tail -1