fix(changelog): cannot generate changelog for first tag
Currently if you try to generate a changelog for the first tag `GIT_TAG_CMD` returns an error which bails out of generating the changelog. This patch address this by parsing all commits from the dawn of time until HEAD for your first tag. It also assumes your first tag is `v0.0.1`.